block, drivers, fs: shrink bi_rw from long to int
We don't need bi_rw to be so large on 64 bit archs, so
reduce it to unsigned int.

+++ include/linux/blk_types.h
@@ -38,7 +38,7 @@ struct bio {
 	unsigned int		bi_flags;	/* status, command, etc */
 	unsigned short		bi_write_hint;
 	int			bi_error;
-	unsigned long		bi_rw;		/* READ/WRITE */
+	unsigned int		bi_rw;		/* READ/WRITE */
 	unsigned short		bi_ioprio;
 
 	struct bvec_iter	bi_iter;
